refactor(phase-48.1): split custom_writer.py from gromacs_writer (no behavior change)
- Move write_custom_molecule_gro_file + write_custom_molecule_top_file to custom_writer.py (569 lines)
- INLINE header write PRESERVED (f'{total_atoms}\n' NO :5d — divergent format, research §3)
- INLINE box vector write PRESERVED (3-value diagonal — NOT _write_gro_box_vectors, 48.1-03 + 48.1-06 finding)
- _write_gro_header NOT called (would impose :5d)
- _write_gro_box_vectors NOT called (would impose 9-value triclinic)
- custom_active 4-writer gate + try/except (OSError, ValueError) cleanup + DeprecationWarning preserved
- _write_top_defaults Format C preserved (comb-rule=2 in helper, not inline)
- gromacs_writer.py re-exports both (1278 → 787 lines, net -491; 49 importable symbols preserved, zero caller changes)
- 14 byte-equivalence (custom path critical — BOTH divergent header AND divergent box vector preserved) + 6 custom grompp + 15 tip4p scan-all-files + 35 MoleculeIndex scan-all-files + 112 tests/test_output regression ALL PASS (1840 total tests, no loss)
